requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
langchain-groq


-e .